+++ src/main/java/core/gui/theme/ho/HOClassicSchema.java
@@ -256,6 +256,8 @@ public class HOClassicSchema extends Schema implements HOIconName, HOColorName, 
 		put(ARROW_CIRCLE, "gui/bilder/Fugue/arrow-circle-double-135.png");
 		put(ARROW_MOVE, "gui/bilder/Fugue/arrow-move-recoloured.png");
 		put(SUBSTITUTION, "gui/bilder/Fugue/arrow-circle-double-135-recoloured.png");
+		// TODO: create new icon for man marking
+		put(MAN_MARKING, "gui/bilder/Fugue/arrow-circle-double-135-recoloured.png");
 		put(IFA_VISITED, "gui/bilder/Fugue/status.png");
 		put(INFORMATION, "gui/bilder/Fugue/information-white.png");
 		put(CHPP, "gui/bilder/chpp.png");