+++ src/main/resources/changelog.md
@@ -46,6 +46,8 @@ If you find a bug, please open an issue on [GitHub](https://github.com/akasolace
 ### Lineup
 
   - [FEAT] Provide better defaults for initial position of dividers.
+  - [FIX] Fix bug in download of man marking match orders [#632]
+  - [FIX] Fix bug in upload of position change match orders [#633]
 
 ### Training